+++ CHANGELOG.md
@@ -2,8 +2,9 @@
 
 ### Unreleased
 
-- Remove gacela file cache. Instead, use in memory cache always.
 - Add (optional) gacela file profiler. By default disabled.
+- Remove gacela file cache. Instead, use in memory cache always.
+- Remove `gacela cache:clear` command.
 
 ### 0.25.0
 #### 2022-09-18